Move Parse_Options to it's own file.
